requirements.txt
PyQt6
numpy
Pillow
opencv-contrib-python
appdirs
pyqtgraph